Blog List Expansion, step 9

+++ blog/backend/controllers/blogs.js
@@ -36,6 +36,18 @@ blogRouter.post('', async (request, response) => {
 
 // DELETE: delete a blog
 blogRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() !== user.id.toString()) {
+        return response.status(401).json({ error: 'unauthorized' })
+    }
+
     const deletedBlog = await Blog.findByIdAndDelete(request.params.id)
 
     if (!deletedBlog) {
